refactor: update selected-posts selectors file name
